arreglado el error al añadir un cliente

+++ src/main/java/com/ad03/ad03/removeCliente.java
@@ -19,11 +19,12 @@ import javax.swing.JOptionPane;
  * @author Manuel
  */
 public class removeCliente extends javax.swing.JDialog {
-    
-    HashMap<String,Integer> clientesMap = new HashMap<>();
+
+    HashMap<String, Integer> clientesMap = new HashMap<>();
 
     /**
      * Creates new form removeTienda
+     *
      * @param parent
      * @param modal
      */
@@ -50,6 +51,7 @@ public class removeCliente extends javax.swing.JDialog {
         jButton2 = new javax.swing.JButton();
 
         setDefaultCloseOperation(javax.swing.WindowConstants.DISPOSE_ON_CLOSE);
+        setTitle("Eliminar Cliente");
 
         jLabel1.setText("Seleccionar Cliente:");
 
@@ -103,13 +105,18 @@ public class removeCliente extends javax.swing.JDialog {
 
     private void jButton2ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jButton2ActionPerformed
         // TODO add your handling code here:
-        String i = (String) this.jComboBoxEliminarCliente.getSelectedItem();
-        int idClienteEliminar = this.clientesMap.get(i);
-        
-            if (JOptionPane.showConfirmDialog(this, "¿Está seguro?", "Advertencia",JOptionPane.OK_CANCEL_OPTION) == JOptionPane.OK_OPTION){
+
+        if (jComboBoxEliminarCliente.getItemCount() != 0) {
+            String i = (String) this.jComboBoxEliminarCliente.getSelectedItem();
+            int idClienteEliminar = this.clientesMap.get(i);
+
+            if (JOptionPane.showConfirmDialog(this, "¿Está seguro?", "Advertencia", JOptionPane.OK_CANCEL_OPTION) == JOptionPane.OK_OPTION) {
                 VentanaPrincipal.eliminarCliente(idClienteEliminar);
                 this.dispose();
-            } 
+            }
+        } else {
+            JOptionPane.showMessageDialog(this, "No existen clientes", "Error", JOptionPane.ERROR_MESSAGE);
+        }
     }//GEN-LAST:event_jButton2ActionPerformed
 
     private void jButton1ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jButton1ActionPerformed
@@ -120,52 +127,12 @@ public class removeCliente extends javax.swing.JDialog {
     /**
      * @param args the command line arguments
      */
-    public static void main(String args[]) {
-        /* Set the Nimbus look and feel */
-        //<editor-fold defaultstate="collapsed" desc=" Look and feel setting code (optional) ">
-        /* If Nimbus (introduced in Java SE 6) is not available, stay with the default look and feel.
-         * For details see http://download.oracle.com/javase/tutorial/uiswing/lookandfeel/plaf.html 
-         */
-        try {
-            for (javax.swing.UIManager.LookAndFeelInfo info : javax.swing.UIManager.getInstalledLookAndFeels()) {
-                if ("Nimbus".equals(info.getName())) {
-                    javax.swing.UIManager.setLookAndFeel(info.getClassName());
-                    break;
-                }
-            }
-        } catch (ClassNotFoundException ex) {
-            java.util.logging.Logger.getLogger(removeCliente.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
-        } catch (InstantiationException ex) {
-            java.util.logging.Logger.getLogger(removeCliente.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
-        } catch (IllegalAccessException ex) {
-            java.util.logging.Logger.getLogger(removeCliente.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
-        } catch (javax.swing.UnsupportedLookAndFeelException ex) {
-            java.util.logging.Logger.getLogger(removeCliente.class.getName()).log(java.util.logging.Level.SEVERE, null, ex);
-        }
-        //</editor-fold>
-        //</editor-fold>
-        //</editor-fold>
-        //</editor-fold>
-
-        /* Create and display the dialog */
-        java.awt.EventQueue.invokeLater(new Runnable() {
-            public void run() {
-                removeCliente dialog = new removeCliente(new javax.swing.JFrame(), true);
-                dialog.addWindowListener(new java.awt.event.WindowAdapter() {
-                    @Override
-                    public void windowClosing(java.awt.event.WindowEvent e) {
-                        System.exit(0);
-                    }
-                });
-                dialog.setVisible(true);
-            }
-        });
-    }
-    
     /**
-     * Con este método establecemos los elementos que se muestran en el desplegable
+     * Con este método establecemos los elementos que se muestran en el
+     * desplegable
      */
-    public void fijarModelo(){
+    public void fijarModelo() {
+
         String sql = "SELECT * FROM Clientes ORDER BY nombre ASC";
 
         ArrayList<String> cadena = new ArrayList<>();
